requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
redis==5.0.1
kafka-python==2.0.2
prometheus-client==0.20.0
structlog==24.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0 